+++ package.json
@@ -12,18 +12,24 @@
     "test": "NODE_OPTIONS=--experimental-vm-modules jest --coverage",
     "badges": "jest-coverage-badges --input '.coverage/coverage-summary.json' --output '.badges'",
     "lint": "eslint --ext js,ts,tsx src",
-    "lint:precommit": "if git diff --name-only HEAD | grep -E '\\.tsx?$'; then npm run lint:fix; else true; fi",
+    "lint:precommit": "if git diff --name-only HEAD | grep -E '\\.tsx?$'; then npm run lint; else true; fi",
     "lint:fix": "eslint --ext js,ts,tsx --fix src",
-    "prettier": "prettier --write \"{src/**/*.{js,jsx,ts,tsx,json,yml,css,scss},travis.yml,*.json}\""
+    "prettier": "prettier --write \"{src/**/*.{js,jsx,ts,tsx,json,yml,css,scss},travis.yml,*.json}\"",
+    "prepare": "husky install",
+    "pre-commit": "yarn run pretty-quick --staged --pattern \"{src/**/*.{js,jsx,ts,tsx,json,yml,css,scss},travis.yml,*.json}\" && npm run lint:precommit"
   },
   "license": "Apache-2.0",
   "author": "Red Hat",
   "devDependencies": {
+    "@patternfly/patternfly": "4.179.1",
+    "@patternfly/react-charts": "6.51.5",
+    "@patternfly/react-core": "4.198.5",
+    "@patternfly/react-styles": "4.48.5",
     "@rollup/plugin-commonjs": "24.0.1",
-    "@rollup/plugin-json": "^6.0.0",
+    "@rollup/plugin-json": "6.0.0",
     "@rollup/plugin-node-resolve": "15.0.1",
-    "@testing-library/dom": "9.0.1",
-    "@testing-library/jest-dom": "5.16.2",
+    "@testing-library/dom": "9.2.0",
+    "@testing-library/jest-dom": "5.16.5",
     "@testing-library/react": "12.1.5",
     "@testing-library/react-hooks": "8.0.1",
     "@testing-library/user-event": "14.4.3",
@@ -31,12 +37,17 @@
     "@types/jest": "29.5.0",
     "@types/react": "17.0.53",
     "@types/react-dom": "17.0.19",
-    "@typescript-eslint/eslint-plugin": "5.56.0",
-    "@typescript-eslint/parser": "5.56.0",
+    "@typescript-eslint/eslint-plugin": "5.57.0",
+    "@typescript-eslint/parser": "5.57.0",
     "@wojtekmaj/enzyme-adapter-react-17": "0.8.0",
+    "axios": "1.3.4",
+    "cytoscape": "3.15.5",
+    "cytoscape-canvas": "3.0.1",
+    "cytoscape-popper": "1.0.7",
+    "deep-freeze": "0.0.1",
     "enzyme": "3.11.0",
     "enzyme-to-json": "3.6.2",
-    "eslint": "8.36.0",
+    "eslint": "8.37.0",
     "eslint-config-prettier": "8.8.0",
     "eslint-plugin-jest": "27.2.1",
     "eslint-plugin-prettier": "4.2.1",
@@ -47,36 +58,22 @@
     "jest-cli": "29.5.0",
     "jest-coverage-badges": "1.1.2",
     "jest-environment-jsdom": "29.5.0",
+    "js-yaml": "3.13.1",
+    "lodash": "4.17.21",
     "postcss": "8.4.21",
-    "prettier": "2.8.6",
+    "prettier": "2.8.7",
     "pretty-quick": "3.1.3",
     "react": "17.0.2",
+    "react-ace": "9.1.3",
     "react-dom": "17.0.2",
     "react-router-dom": "5.2.0",
-    "rollup": "3.20.0",
+    "rollup": "3.20.2",
     "rollup-plugin-peer-deps-external": "2.2.4",
     "rollup-plugin-postcss": "4.0.2",
     "rollup-plugin-typescript2": "0.34.1",
     "ts-jest": "29.0.5",
     "typescript": "4.9.5",
     "typestyle": "2.4.0"
-  },
-  "dependencies": {
-    "@patternfly/patternfly": "4.179.1",
-    "@patternfly/react-charts": "6.51.5",
-    "@patternfly/react-core": "4.198.5",
-    "@patternfly/react-styles": "4.48.5",
-    "@patternfly/react-table": "4.67.5",
-    "@patternfly/react-tokens": "4.50.5",
-    "axios": "0.21.4",
-    "cytoscape": "3.15.5",
-    "cytoscape-canvas": "3.0.1",
-    "cytoscape-popper": "1.0.7",
-    "deep-freeze": "0.0.1",
-    "js-yaml": "3.13.1",
-    "lodash": "^4.17.21",
-    "react-ace": "9.1.3",
-    "seedrandom": "3.0.1"
   },
   "peerDependencies": {
     "react": "^17.0.0",
@@ -89,10 +86,5 @@
     "node": ">=14.17.0",
     "npm": ">=6.0.0 ",
     "yarn": ">=1.0.0 "
-  },
-  "husky": {
-    "hooks": {
-      "pre-commit": "yarn run pretty-quick --staged --pattern \"{src/**/*.{js,jsx,ts,tsx,json,yml,css,scss},travis.yml,*.json}\" && npm run lint:precommit"
-    }
   }
 }
