+++ src/types/index.ts
@@ -9,8 +9,9 @@ export * from './ErrorMsg';
 export * from './Filters';
 export * from './GrafanaInfo';
 export * from './Graph';
-export * from './HealthAnnotation';
 export * from './Health';
+export * from './HealthAnnotation';
+export * from './HealthStatus';
 export * from './IstioConfigDetails';
 export * from './IstioConfigList';
 export * from './IstioObjects';
