+++ tsconfig.json
@@ -9,7 +9,6 @@
     "allowJs": true,
     "skipLibCheck": true,
     "esModuleInterop": true,
-    "allowSyntheticDefaultImports": true,
     "strict": true,
     "forceConsistentCasingInFileNames": true,
     "module": "esnext",
@@ -17,7 +16,7 @@
     "resolveJsonModule": true,
     "isolatedModules": true,
     "noEmit": true,
-    "jsx": "react",  
+    "jsx": "react",
     "sourceMap": true,
     "noImplicitReturns": true,
     "noImplicitThis": true,
@@ -31,5 +30,3 @@
   "include": ["src"],
   "exclude": ["node_modules", "lib"]
 }
-
-
